Avoid divide by zero in SQ1 servo SA tuning.
Previously, fits were broken if the I gain was zero, as this would introduce
NaNs in the data and produce runtime warnings. Now, the P gain is used instead
if the I gain is zero, and 1 is used if both are, avoiding a divide by zero
and fixing the fits.

This was a problem for BICEP3, since it runs with just P gain (a similar, but
less generic, change has been a local modification since 2015).